test(diagnostics): split the port cases out of tests.rs
The port-exposure tests took tests.rs to 502 code lines, two over the limit.
They move to a `#[path]` sibling, which is the convention this tree already
uses — tests.rs itself carries a header saying it was split out of mod.rs for
the same reason.

The helper needs `super::super::collect` now: the module sits one level
deeper than before, so `super` is `tests` rather than `diagnostics`. Same 68
tests pass, and nothing under internal/ exceeds 500.
